Add `match` property to `expect_out`
`match` is analogous to the expect_out pattern matches in TCL Expect.
Instead of expect_out.[0|1|n], we'll have a single `match` property that
will contain the string we matched on if the expectation was a String,
or the result object of the successful test if the expectation was a
RegExp. Glob support is planned as well.

Also add ViaSat, Inc. to the license copyright.

In addition, some `eslint` errors were added by me; `pre-git` was
uninstalled in my local mirror and I didn't notice, so my commits
were not being linted.
This commit fixes all `eslint` errors introduced.

Resolves #11
See also #18