Add support for `timeout` and `eof` expecatations
This commit will allow users to expect timeout and eof on the session.
Also added is a 'final' callback for the expect function, which is
called from a handler function or on *unhandled* EOF/TIMEOUT.

In order to properly support this, a new stream is added that, when
paused, caches its messages to a buffer.  STDOUT/ERR is piped to this
stream, which is paused when we're not actively expecting.

Updates tests to account for new functionality.

In addition, this commit will add a "Getting Started" section to the
README that better explains how to get started with spectcl.

Comments and minor code cleanup also included.

Resolves #15