Add callbacks to send and sendEof
This commit adds optional callbacks to send and sendEof functions to
better support "fire-and-forget" operations.  These are operations where
the issuer does not want to wait around to examine the result of the
command.  Normally such a callback might not be used because a send
would be immediately followed by an expect, but there are times when the
user wants to return only after the command has been written to the
session channel.

Docs are updated to reflect the new callbacks.
Tests are also added to vet the callback.

Resolves #69